fix formatting

+++ trunk/core/malloc.cpp
@@ -1722,9 +1722,9 @@ static void internal_malloc_stats(mstate m) {
     }
 
 #ifdef LACKS_STDIO_H
-    console("max system bytes = %10lu\n", (unsigned long)(maxfp));
-    console("system bytes     = %10lu\n", (unsigned long)(fp));
-    console("in use bytes     = %10lu\n", (unsigned long)(used));
+    console("max system bytes = %u\n", (uint)maxfp);
+    console("system bytes     = %u\n", (uint)fp);
+    console("in use bytes     = %u\n", (uint)used);
 #else
     fprintf(stderr, "max system bytes = %10lu\n", (unsigned long)(maxfp));
     fprintf(stderr, "system bytes     = %10lu\n", (unsigned long)(fp));
